src/agent/runtime/requirements.txt
azure-ai-agentserver-agentframework==1.0.0b17
azure-identity==1.25.3
httpx==0.28.1
python-dotenv==1.2.2
